Fixed PNG decoding once again.
git-svn-id: http://svn.purei.org/framework/trunk@300 6a6d099a-6a11-0410-877e-d5a07a98cbd2

+++ include/PNG.h
@@ -35,10 +35,10 @@ namespace Framework
 
 		CBitmap				DoRead(CStream&);
 		void				UncompressIDAT();
-		void				SubFilter(uint8*, size_t, unsigned int, unsigned int);
-		void				UpFilter(uint8*, size_t, unsigned int, unsigned int);
-		void				AverageFilter(uint8*, size_t, unsigned int, unsigned int);
-		void				PaethFilter(uint8*, size_t, unsigned int, unsigned int);
+		void				SubFilter(uint8*, int, unsigned int, unsigned int);
+		void				UpFilter(uint8*, int, unsigned int, unsigned int);
+		void				AverageFilter(uint8*, int, unsigned int, unsigned int);
+		void				PaethFilter(uint8*, int, unsigned int, unsigned int);
 		uint8				PaethPredictor(uint8, uint8, uint8);
 		CBitmap				CreateBitmap();
 		CBitmap				ConvertTo32(const CBitmap&) const;
