+++ src/PNG.cpp
@@ -105,11 +105,11 @@ void CPNG::UncompressIDAT()
 	}
 }
 
-void CPNG::SubFilter(uint8* pRaw, size_t nPitch, unsigned int nScanline, unsigned int nBPP)
+void CPNG::SubFilter(uint8* pRaw, int nPitch, unsigned int nScanline, unsigned int nBPP)
 {
-	size_t nDist = (nBPP / 8);
+	int nDist = (nBPP / 8);
 
-	for(int i = 0; i < static_cast<int>(nPitch); i++)
+	for(int i = 0; i < nPitch; i++)
 	{
 		uint8 nPredictor = 0;
 		if((i - nDist) < 0)
@@ -124,9 +124,9 @@ void CPNG::SubFilter(uint8* pRaw, size_t nPitch, unsigned int nScanline, unsigne
 	}
 }
 
-void CPNG::UpFilter(uint8* pRaw, size_t nPitch, unsigned int nScanline, unsigned int nBPP)
+void CPNG::UpFilter(uint8* pRaw, int nPitch, unsigned int nScanline, unsigned int nBPP)
 {
-	for(int i = 0; i < static_cast<int>(nPitch); i++)
+	for(int i = 0; i < nPitch; i++)
 	{
 		uint8 nPredictor = 0;
 		if(nScanline == 0)
@@ -141,11 +141,11 @@ void CPNG::UpFilter(uint8* pRaw, size_t nPitch, unsigned int nScanline, unsigned
 	}
 }
 
-void CPNG::AverageFilter(uint8* pRaw, size_t nPitch, unsigned int nScanline, unsigned int nBPP)
+void CPNG::AverageFilter(uint8* pRaw, int nPitch, unsigned int nScanline, unsigned int nBPP)
 {
-	size_t nDist = (nBPP / 8);
+	int nDist = (nBPP / 8);
 
-	for(int i = 0; i < static_cast<int>(nPitch); i++)
+	for(int i = 0; i < nPitch; i++)
 	{
 		unsigned int nAbove = 0;
 		if(nScanline == 0)
@@ -184,9 +184,9 @@ uint8 CPNG::PaethPredictor(uint8 a, uint8 b, uint8 c)
 	return c;
 }
 
-void CPNG::PaethFilter(uint8* pRaw, size_t nPitch, unsigned int nScanline, unsigned int nBPP)
+void CPNG::PaethFilter(uint8* pRaw, int nPitch, unsigned int nScanline, unsigned int nBPP)
 {
-	size_t nDist = (nBPP / 8);
+	int nDist = (nBPP / 8);
 
 	for(int i = 0; i < static_cast<int>(nPitch); i++)
 	{
